Changes due to comments on the first iteration of this task from Kyrylo Haiduk
