+++ src/reducer.ts
@@ -1,13 +1,8 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import React, { useEffect, useReducer } from 'react';
-
-import * as todoService from '../../api/todos';
-
-import { Filter } from '../../types/Filter';
-import { Todo } from '../../types/Todo';
-import { TodoError } from '../../types/TodoError';
-
-export const USER_ID = 11722;
+import { State } from './types/State';
+import { Todo } from './types/Todo';
+import { Filter } from './types/Filter';
+import { TodoError } from './types/TodoError';
 
 enum ReducerActions {
   LoadTodos = 'loadTodos',
@@ -24,24 +19,18 @@ enum ReducerActions {
   ClearTodoErrors = 'clearTodoErrors',
 }
 
-interface Action {
+export interface Action {
   type: ReducerActions;
   payload?: any;
 }
 
-interface State {
-  initialTodos: Todo[];
-  selectedFilter: Filter;
-  tempTodo: Todo | null;
-  loadingItemsId: number[];
-  isSubmitting: boolean;
-  isUpdating: boolean;
-  isDeleting: boolean;
-  isEditing: boolean;
-  todoError: TodoError | null;
+interface UpdateState {
+  add?: Todo;
+  update?: Todo;
+  delete?: number;
 }
 
-function reducer(state: State, action: Action): State {
+export function reducer(state: State, action: Action): State {
   switch (action.type) {
     case ReducerActions.LoadTodos:
       return {
@@ -58,17 +47,18 @@ function reducer(state: State, action: Action): State {
 
       if (action.payload.update) {
         const updatedTodos = [...initialTodos];
-        const index = updatedTodos.findIndex(todo => (
-          todo.id === action.payload.update.id
-        ));
+        const index = updatedTodos.findIndex(
+          (todo) => todo.id === action.payload.update.id,
+        );
 
         updatedTodos.splice(index, 1, action.payload.update);
         initialTodos = updatedTodos;
       }
 
       if (action.payload.delete) {
-        initialTodos = initialTodos
-          .filter(todo => action.payload.delete !== todo.id);
+        initialTodos = initialTodos.filter(
+          (todo) => action.payload.delete !== todo.id,
+        );
       }
 
       return {
@@ -134,7 +124,7 @@ function reducer(state: State, action: Action): State {
     case ReducerActions.ClearTodoErrors:
       return {
         ...state,
-        todoError: null,
+        todoError: TodoError.NoProblem,
       };
 
     default:
@@ -142,77 +132,35 @@ function reducer(state: State, action: Action): State {
   }
 }
 
-interface UpdateState {
-  add?: Todo,
-  update?: Todo,
-  delete?: number,
-}
-
 export const actionCreator = {
   load: (payload: Todo[]) => ({ type: ReducerActions.LoadTodos, payload }),
   updateTodos: (payload: UpdateState) => ({
-    type: ReducerActions.UpdateTodos, payload,
+    type: ReducerActions.UpdateTodos,
+    payload,
   }),
   selectFilter: (payload: Filter) => ({
-    type: ReducerActions.SelectFilter, payload,
+    type: ReducerActions.SelectFilter,
+    payload,
   }),
   addTempTodo: (payload: Todo | null) => ({
-    type: ReducerActions.AddTempTodo, payload,
+    type: ReducerActions.AddTempTodo,
+    payload,
   }),
   addLoadingItemId: (payload: number) => ({
-    type: ReducerActions.AddLoadingItemId, payload,
+    type: ReducerActions.AddLoadingItemId,
+    payload,
   }),
   clearLoadingItemsId: () => ({ type: ReducerActions.ClearLoadingItemsId }),
   toggleSubmitting: () => ({ type: ReducerActions.ToggleSubmitting }),
   toggleUpdating: () => ({ type: ReducerActions.ToggleUpdating }),
   toggleDeleting: () => ({ type: ReducerActions.ToggleDeleting }),
   toggleEditing: (payload: boolean) => ({
-    type: ReducerActions.ToggleEditing, payload,
+    type: ReducerActions.ToggleEditing,
+    payload,
   }),
   addError: (payload: TodoError) => ({
-    type: ReducerActions.AddTodoError, payload,
+    type: ReducerActions.AddTodoError,
+    payload,
   }),
   clearError: () => ({ type: ReducerActions.ClearTodoErrors }),
-};
-
-const initialState: State = {
-  initialTodos: [],
-  selectedFilter: Filter.All,
-  tempTodo: null,
-  loadingItemsId: [],
-  isSubmitting: false,
-  isUpdating: false,
-  isDeleting: false,
-  isEditing: false,
-  todoError: null,
-};
-
-export const StateContext = React.createContext(initialState);
-export const DispatchContext = React.createContext<React.Dispatch<Action>>(
-  () => {},
-);
-
-type Props = {
-  children: React.ReactNode;
-};
-
-export const TodoAppContext: React.FC<Props> = ({ children }) => {
-  const [state, dispatch] = useReducer(reducer, initialState);
-
-  useEffect(() => {
-    dispatch(actionCreator.clearError());
-    todoService.getTodos(USER_ID)
-      .then(todos => dispatch(actionCreator.load(todos)))
-      .catch(() => {
-        dispatch(actionCreator.addError(TodoError.ErrorLoad));
-      });
-  }, []);
-
-  return (
-    <DispatchContext.Provider value={dispatch}>
-      <StateContext.Provider value={state}>
-        {children}
-      </StateContext.Provider>
-    </DispatchContext.Provider>
-  );
 };
